LteHarqUnitTx: fix: add missing int cast to acid for writing it to stream
Otherwise a NUL characher byte (\x0) being written to log instead of "0",
because its type is char. This was done everywhere else in the code,
but was missing here.

+++ src/simu5g/stack/mac/buffer/harq/LteHarqUnitTx.cc
@@ -73,10 +73,10 @@ void LteHarqUnitTx::insertPdu(Packet *pkt)
 void LteHarqUnitTx::markSelected()
 {
     EV << NOW << " LteHarqUnitTx::markSelected trying to select buffer "
-       << acid_ << " codeword " << cw_ << " for transmission " << endl;
+       << (unsigned int)acid_ << " codeword " << cw_ << " for transmission " << endl;
 
     if (!(this->isReady()))
-        throw cRuntimeError("ERROR acid %d codeword %d trying to select for transmission an empty buffer", acid_, cw_);
+        throw cRuntimeError("ERROR acid %d codeword %d trying to select for transmission an empty buffer", (unsigned int)acid_, cw_);
 
     status_ = TXHARQ_PDU_SELECTED;
 }
